修复(项目): 删除项目时级联清理角色
项目存在草稿或已发布角色时无法从正常界面完成删除。

删除流程在项目行锁内清理全部角色，再删除项目并尽力清理媒体。

单角色删除与项目删除共用同一套媒体键提取规则。
