+++ backend/packages/app/src/windup_app/web/api/character.py
@@ -11,15 +11,16 @@ from windup_common.enums.biz_code import BizCode
 from windup_common.enums.character import CharacterStatus
 from windup_common.exceptions import BizException
 from windup_common.result import ListResponse, Response
-from windup_framework.config.storage import settings as storage_settings
 from windup_framework.db import get_session
 
+from windup_app.server.character.cleanup import (
+    extract_object_keys as _extract_object_keys,
+)
 from windup_app.server.character.model import Character, CharacterData
 from windup_app.server.character.direction_validation import (
     validate_character_directions,
 )
 from windup_app.server.character.service import service as character_service
-from windup_app.server.media.service import card_thumbnail_key
 from windup_app.server.media.service import service as media_service
 from windup_app.server.project.model import Project
 from windup_app.server.project.service import service as project_service
@@ -78,52 +79,6 @@ class CharacterSummaryOut(BaseModel):
     outfit_name: str | None
     outfit_count: int
     action_count: int
-
-
-# ── 辅助函数 ─────────────────────────────────────────────────────────────────
-
-
-def _extract_object_keys(character: Character) -> list[str]:
-    """从角色中提取所有对象存储 key,用于删除时清理资源。
-
-    URL 格式: ``{download_base}/{object_key}``
-    """
-    prefix = storage_settings.download_base + "/"
-    keys: list[str] = []
-    seen: set[str] = set()
-
-    def add_url(url: str | None, *, include_thumbnail: bool = False) -> None:
-        if not url or not url.startswith(prefix):
-            return
-        key = url[len(prefix) :]
-        if key not in seen:
-            seen.add(key)
-            keys.append(key)
-        if include_thumbnail:
-            thumbnail_key = card_thumbnail_key(key)
-            if thumbnail_key not in seen:
-                seen.add(thumbnail_key)
-                keys.append(thumbnail_key)
-
-    # 参考图
-    add_url(character.reference_image_url, include_thumbnail=True)
-
-    # character_data 内的 URL
-    data = character.character_data or {}
-    for template in data.get("templates", []):
-        add_url(template.get("image_url"))
-    for outfit in data.get("outfits", []):
-        add_url(outfit.get("preview_url"), include_thumbnail=True)
-        for action in outfit.get("actions", []):
-            frame_groups = [action.get("frames", [])]
-            frame_groups.extend(
-                sequence.get("frames", []) for sequence in action.get("sequences", [])
-            )
-            for frames in frame_groups:
-                for frame in frames:
-                    add_url(frame.get("image_url"))
-
-    return keys
 
 
 # ── 归属校验 ─────────────────────────────────────────────────────────────────
